Update README.md
A random forest classifier worked better in predicting the default status of customers. The date-time columns were treated by split and divide to find the relation with target variable and found that the year column is highly correlated to target rather than month and date. Used VIF to find the degree of multicollinearity and removed the features with the sensible decision.